Doctor hash table
There are a few bugs with the availability array that I can't iron out right now. Now that I'm considering it, it'd probably be better to use something else, but that's what's there for now.
I left the hash function blank since I couldn't remember how we said we wanted to organize the table. I'll go back in and add it once we figure that out.